Moved decorator to separate file
